requirements.txt
Flask
google-generativeai
requests
python-dotenv
gunicorn
protobuf==4.25.3